+++ hoomd/ParticleGroup.cc
@@ -18,180 +18,6 @@ using namespace std;
 namespace py = pybind11;
 
 /*! \param sysdef System the particles are to be selected from
-*/
-ParticleFilter::ParticleFilter()
-    {
-    }
-
-/*! \param tag Tag of the particle to check
-    \returns true if the particle is selected
-    \returns false if it is not
-*/
-std::vector<unsigned int> ParticleFilter::getSelectedTags(std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    // base class doesn't do anything useful
-    return std::vector<unsigned int>();
-    }
-
-ParticleFilterAll::ParticleFilterAll()
-    : ParticleFilter()
-    { }
-
-/*! \param tag All of the particle to check
-    \returns true if particle is local
-*/
-std::vector<unsigned int> ParticleFilterAll::getSelectedTags(std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    std::vector<unsigned int> member_tags;
-    auto pdata = sysdef->getParticleData();
-
-    // loop through local particles and select those that match selection criterion
-    ArrayHandle<unsigned int> h_tag(pdata->getTags(), access_location::host, access_mode::read);
-    for (unsigned int idx = 0; idx < pdata->getN(); ++idx)
-        {
-        unsigned int tag = h_tag.data[idx];
-        member_tags.push_back(tag);
-        }
-
-    return member_tags;
-    }
-
-//////////////////////////////////////////////////////////////////////////////
-// ParticleFilterTags
-
-/** Args:
-        tags: numpy array of tags to select
-*/
-ParticleFilterTags::ParticleFilterTags(std::vector<unsigned int> tags)
-    : ParticleFilter(), m_tags(tags)
-    {
-    }
-
-/** Args:
-        tags: numpy array of tags to select
-*/
-ParticleFilterTags::ParticleFilterTags(pybind11::array_t<unsigned int, pybind11::array::c_style> tags)
-    : ParticleFilter()
-    {
-	unsigned int* tags_ptr = (unsigned int*)m_tags.data();
-    m_tags.assign(tags_ptr, tags_ptr+m_tags.size());
-    }
-
-/*! \param tag Tag of the particle to check
-    \returns true if \a m_tag_min <= \a tag <= \a m_tag_max
-*/
-std::vector<unsigned int> ParticleFilterTags::getSelectedTags(std::shared_ptr<SystemDefinition> sysdef) const
-    {
-	return m_tags;
-    }
-
-/*! \param typ_min Minimum type id to select (inclusive)
-    \param typ_max Maximum type id to select (inclusive)
-*/
-ParticleFilterType::ParticleFilterType(
-        std::unordered_set<std::string> included_types)
-    : ParticleFilter(), m_types(included_types)
-    {
-    }
-
-/*! \param tag Tag of the particle to check
-    \returns true if the type of particle \a tag is in the inclusive range [ \a m_typ_min, \a m_typ_max ]
-*/
-std::vector<unsigned int> ParticleFilterType::getSelectedTags(std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    std::vector<unsigned int> member_tags;
-    auto pdata = sysdef->getParticleData();
-
-    // loop through local particles and select those that match selection criterion
-    ArrayHandle<unsigned int> h_tag(pdata->getTags(), access_location::host, access_mode::read);
-    ArrayHandle<Scalar4> h_postype(pdata->getPositions(), access_location::host, access_mode::read);
-    std::unordered_set<unsigned int> types(m_types.size());
-    for (auto type_str: m_types)
-        {
-        types.insert(pdata->getTypeByName(type_str));
-        }
-
-    for (unsigned int idx = 0; idx < pdata->getN(); ++idx)
-        {
-        unsigned int tag = h_tag.data[idx];
-        unsigned int typ = __scalar_as_int(h_postype.data[idx].w);
-        if (types.count(typ))
-            member_tags.push_back(tag);
-        }
-    return member_tags;
-    }
-
-// Set operations filter classes. Each takes two filters and performs the
-// correct set operation on the individual filters. This supports arbitrary
-// nesting.
-
-ParticleFilterSetDifference::ParticleFilterSetDifference(
-        std::shared_ptr<ParticleFilter> f, std::shared_ptr<ParticleFilter> g):
-    m_f(f), m_g(g){}
-
-std::vector<unsigned int> ParticleFilterSetDifference::getSelectedTags(
-    std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    // Get tags for f() and g()
-    auto X = m_f->getSelectedTags(sysdef);
-    std::sort(X.begin(), X.end());
-
-    auto Y = m_g->getSelectedTags(sysdef);
-    std::sort(Y.begin(), Y.end());
-
-    // Create vector and get set difference
-    auto tags = std::vector<unsigned int>(X.size());
-    auto it = set_difference(X.begin(), X.end(), Y.begin(), Y.end(),
-                             tags.begin());
-    tags.resize(it - tags.begin());
-    return tags;
-    }
-
-ParticleFilterIntersection::ParticleFilterIntersection(
-        std::shared_ptr<ParticleFilter> f, std::shared_ptr<ParticleFilter> g):
-    m_f(f), m_g(g){}
-
-std::vector<unsigned int> ParticleFilterIntersection::getSelectedTags(
-    std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    // Get tags for f() and g() as sets
-    auto X = m_f->getSelectedTags(sysdef);
-    std::sort(X.begin(), X.end());
-
-    auto Y = m_g->getSelectedTags(sysdef);
-    std::sort(Y.begin(), Y.end());
-
-    // Create vector and get intersection
-    auto tags = std::vector<unsigned int>(std::min(X.size(), Y.size()));
-    auto it = set_intersection(X.begin(), X.end(), Y.begin(), Y.end(),
-                             tags.begin());
-    tags.resize(it - tags.begin());
-    return tags;
-    }
-
-ParticleFilterUnion::ParticleFilterUnion(
-        std::shared_ptr<ParticleFilter> f, std::shared_ptr<ParticleFilter> g):
-    m_f(f), m_g(g){}
-
-std::vector<unsigned int> ParticleFilterUnion::getSelectedTags(
-    std::shared_ptr<SystemDefinition> sysdef) const
-    {
-    // Get tags for f() and g() as sets
-    auto X = m_f->getSelectedTags(sysdef);
-    std::sort(X.begin(), X.end());
-
-    auto Y = m_g->getSelectedTags(sysdef);
-    std::sort(Y.begin(), Y.end());
-
-    // Create vector and get union
-    auto tags = std::vector<unsigned int>(X.size() + Y.size());
-    auto it = set_union(X.begin(), X.end(), Y.begin(), Y.end(),
-                             tags.begin());
-    tags.resize(it - tags.begin());
-    return tags;
-    }
-
-/*! \param sysdef System the particles are to be selected from
     \param rigid true selects particles that are in bodies, false selects particles that are not part of a body
 */
 ParticleFilterBody::ParticleFilterBody(bool body)
@@ -961,63 +787,4 @@ void export_ParticleGroup(py::module& m)
             .def("groupDifference", &ParticleGroup::groupDifference)
             .def("updateMemberTags", &ParticleGroup::updateMemberTags)
             ;
-
-    py::class_<ParticleFilter, std::shared_ptr<ParticleFilter> >(m,"ParticleFilter")
-            .def(py::init< >())
-            .def("_get_selected_tags", &ParticleFilter::getSelectedTags)
-            ;
-
-    py::class_<ParticleFilterAll, ParticleFilter, std::shared_ptr<ParticleFilterAll> >(m,"ParticleFilterAll")
-            .def(py::init< >())
-        ;
-
-    py::class_<ParticleFilterTags, ParticleFilter, std::shared_ptr<ParticleFilterTags> >(m,"ParticleFilterTags")
-			.def(py::init<pybind11::array_t<unsigned int, pybind11::array::c_style> >())
-        ;
-
-    py::class_<ParticleFilterType, ParticleFilter,
-               std::shared_ptr<ParticleFilterType> >(m,"ParticleFilterType")
-            .def(py::init<std::unordered_set<std::string>>())
-        ;
-
-    py::class_<ParticleFilterSetDifference, ParticleFilter,
-               std::shared_ptr<ParticleFilterSetDifference>
-                   >(m, "ParticleFilterSetDifference")
-            .def(py::init<std::shared_ptr<ParticleFilter>,
-                          std::shared_ptr<ParticleFilter> >())
-        ;
-
-    py::class_<ParticleFilterIntersection, ParticleFilter,
-               std::shared_ptr<ParticleFilterIntersection>
-                   >(m, "ParticleFilterIntersection")
-            .def(py::init<std::shared_ptr<ParticleFilter>,
-                          std::shared_ptr<ParticleFilter> >())
-        ;
-
-    py::class_<ParticleFilterUnion, ParticleFilter,
-               std::shared_ptr<ParticleFilterUnion> >(m, "ParticleFilterUnion")
-            .def(py::init<std::shared_ptr<ParticleFilter>,
-                          std::shared_ptr<ParticleFilter> >())
-        ;
-
-    py::class_<ParticleFilterRigid, ParticleFilter, std::shared_ptr<ParticleFilterRigid> >(m,"ParticleFilterRigid")
-            .def(py::init< bool >())
-        ;
-
-    py::class_<ParticleFilterBody, ParticleFilter, std::shared_ptr<ParticleFilterBody> >(m,"ParticleFilterBody")
-
-            .def(py::init< bool >())
-        ;
-
-    py::class_<ParticleFilterFloppy, ParticleFilter, std::shared_ptr<ParticleFilterFloppy> >(m,"ParticleFilterFloppy")
-            .def(py::init< bool >())
-        ;
-
-    py::class_<ParticleFilterCuboid, ParticleFilter, std::shared_ptr<ParticleFilterCuboid> >(m,"ParticleFilterCuboid")
-            .def(py::init< Scalar3, Scalar3 >())
-        ;
-
-    py::class_<ParticleFilterRigidCenter, ParticleFilter, std::shared_ptr<ParticleFilterRigidCenter> >(m,"ParticleFilterRigidCenter")
-            .def(py::init< >())
-        ;
     }
