+++ hoomd/filter/export_filters.h
@@ -1,0 +1,7 @@
+#ifndef __EXPORT_FILTERS_H__
+#define __EXPORT_FILTERS_H__
+
+#include <pybind11/pybind11.h>
+
+void export_ParticleFilters(pybind11::module& m);
+#endif
